// Angular/project/src/app/modelos/crear-oferta-datgen.interface.ts
export interface datGenI {
  id_oferta: Number | null | undefined;
  derecho_tipo: String | null | undefined;
  tipo_inmueble: String | null | undefined;
  si_valor_incluye_anexidades: String | null | undefined;
  fecha: Date | null | undefined;
  tiempo_oferta_mercado: Number | null | undefined;
  proyecto_inmobiliario: String | null | undefined;
  proyecto_descripcion: Text | unknown;
}